pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "dlpbot"
dynamic = ["version"]
description = "A Discord moderation bot for yt-dlp"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Simon Sawicki", email = "contact@grub4k.xyz" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "discord-py>=2.4.0",
]

[project.scripts]
dlpbot = "dlpbot.__main__:main"

[project.urls]
Documentation = "https://github.com/Grub4K/dlpbot#readme"
Issues = "https://github.com/Grub4K/dlpbot/issues"
Source = "https://github.com/Grub4K/dlpbot"
Funding = "https://github.com/sponsors/Grub4K"

[tool.hatch.envs.default]
installer = "uv"
dependencies = ["pre-commit"]

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "src/dlpbot/_version.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/dlpbot}"

[tool.hatch.envs.hatch-test.scripts]
run = "echo No tests implemented && exit 1"
run-cov = "echo No tests implemented && exit 1"

[tool.hatch.envs.hatch-static-analysis]
installer = "uv"
dependencies = ["ruff==0.8.*"]
config-path = "pyproject.toml"

[tool.ruff.lint]
select = [
  "C4",
  "E",
  "F",
  "I",
  "PLC",
  "PLE",
  "PLW",
  "PYI",
  "RET",
  "RUF",
  "SIM",
  "TD",
  "TID",
  "W",
]
ignore = [
  "TD003",
  "E402",
  "E501",
  "PLR09",
]

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]
force-single-line = true

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
